+++ src/Parser.php
@@ -110,6 +110,14 @@ class Parser extends Core
             'field' => 'end_options',
         ),
 
+        'INTERSECT' => array(
+            'class' => 'PhpMyAdmin\\SqlParser\\Components\\UnionKeyword',
+            'field' => 'union',
+        ),
+        'EXCEPT' => array(
+            'class' => 'PhpMyAdmin\\SqlParser\\Components\\UnionKeyword',
+            'field' => 'union',
+        ),
         'UNION' => array(
             'class' => 'PhpMyAdmin\\SqlParser\\Components\\UnionKeyword',
             'field' => 'union',
@@ -427,7 +435,12 @@ class Parser extends Core
                 continue;
             }
 
-            if (($token->keyword === 'UNION') || ($token->keyword === 'UNION ALL') || ($token->keyword === 'UNION DISTINCT')) {
+            if (($token->keyword === 'UNION') ||
+                    ($token->keyword === 'UNION ALL') ||
+                    ($token->keyword === 'UNION DISTINCT') ||
+                    ($token->keyword === 'EXCEPT') ||
+                    ($token->keyword === 'INTERSECT')
+            ) {
                 $unionType = $token->keyword;
                 continue;
             }